+++ .gitignore
@@ -37,5 +37,3 @@ out/
 .vscode/
 gradlew
 gradlew.bat
-/logs/
-/src/main/resources/vouchers.csv
