+++ src/main/java/org/prgrms/kdt/io/ConsoleManager.java
@@ -1,6 +1,5 @@
 package org.prgrms.kdt.io;
 
-import org.prgrms.kdt.customer.Customer;
 import org.prgrms.kdt.voucher.Voucher;
 import org.springframework.stereotype.Service;
 
@@ -13,8 +12,7 @@ public class ConsoleManager {
             "=== Voucher Program ===" + System.lineSeparator() +
             "Type exit to exit the program." + System.lineSeparator() +
             "Type create to create a new voucher." + System.lineSeparator() +
-            "Type list to list all vouchers" + System.lineSeparator() +
-            "Type black to show black customer list" + System.lineSeparator();
+            "Type list to list all vouchers" + System.lineSeparator();
     public static final String VOUCHER_TYPE_PROMPT = System.lineSeparator() + "Enter voucher type. (fixed or percent)";
     public static final String VOUCHER_AMOUNT_PROMPT = "Enter a value. (fixed - 0~ | percent - 0~100)";
 
@@ -52,14 +50,6 @@ public class ConsoleManager {
         output.printText(
                 vouchers.stream()
                         .map(Voucher::toString)
-                        .collect(Collectors.joining("\n"))
-        );
-    }
-
-    public void printBlackList(List<Customer> blacklist) {
-        output.printText(
-                blacklist.stream()
-                        .map(Customer::toString)
                         .collect(Collectors.joining("\n"))
         );
     }
