+++ src/main/java/org/prgrms/kdt/voucher/PercentDiscountVoucher.java
@@ -3,7 +3,6 @@ package org.prgrms.kdt.voucher;
 public class PercentDiscountVoucher implements Voucher {
     private static final Long PERCENT_DISCOUNT_VOUCHER_MIN_VALUE = 0L;
     private static final Long PERCENT_DISCOUNT_VOUCHER_MAX_VALUE = 100L;
-    private static final String TYPE = "PERCENT";
 
     private final long id;
     private final VoucherAmount amount;
@@ -27,16 +26,6 @@ public class PercentDiscountVoucher implements Voucher {
     @Override
     public long getId() {
         return id;
-    }
-
-    @Override
-    public String getType() {
-        return TYPE;
-    }
-
-    @Override
-    public String getAmount() {
-        return String.valueOf(amount.getValue());
     }
 
     @Override
