+++ src/main/java/org/prgrms/kdt/voucher/Voucher.java
@@ -10,8 +10,4 @@ public interface Voucher {
     }
 
     long getId();
-
-    String getType();
-
-    String getAmount();
 }
