--- src/main/resources/customer_blacklist.csv
@@ -1,27 +1,0 @@
-1, 바퀴벌레
-2, 곱등이
-3, 돈벌레
-4, 연가시
-5, 미세먼지
-6, 공기청정기
-7, 기차
-8, 차표
-9, 표범
-10, 범인
-11, 인간
-12, 간장
-13, 장난꾸러기
-14, 기차
-15, 차표
-16, 표범
-17, 범인
-18, 인간
-19, 간장
-20, 장난꾸러기
-21, 기차
-22, 차표
-23, 표범
-24, 범인
-25, 인간
-26, 간장
-27, 장난꾸러기
