--- src/test/java/org/prgrms/kdt/voucher/FileVoucherManagerTest.java
@@ -1,7 +1,0 @@
-package org.prgrms.kdt.voucher;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class FileVoucherManagerTest {
-    
-}
